requirements.txt
astroid==2.4.2
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
python-slugify==4.0.1
six==1.15.0
text-unidecode==1.3
toml==0.10.2
wrapt==1.12.1
